Handle loading of account in base activity
Eventual goal is to ensure that any DroidBooruAccountActivity subclass has
access to an account as soon as the subclass' lifecycle is triggered
(particularly after first launch when the user is required to pick an
account). Currently the activity is allowed to run in the background of the
account picker. A better solution is desired.