Reorganize activities
The app can now be entered from either the Gallery or UploadFiles activities,
depending on what function the user activates first. If an account hasn't been
selected, the user is prompted immediately.

Backend.getInstance() now requires a Context, but this allows it to be
instantiated at any time if necessary, which is important for the above app
flow changes.